Replace isize with i32

+++ src/builder.rs
@@ -1,7 +1,7 @@
 use sudoku::Puzzle;
 use solver::Solver;
 
-pub fn build_complete_puzzle() -> Vec<Vec<isize>> {
+pub fn build_complete_puzzle() -> Vec<Vec<i32>> {
     let grid = vec![vec![0;9];9];
     let puzzle = Puzzle::new(grid);
     let mut solver = Solver { puzzle };
